move respondToButton fn to its own file
