+++ src/lib/util/respondToButton.ts
@@ -1,0 +1,19 @@
+import { InteractionResponseType, MessageFlags, Routes } from 'mahoji';
+
+export async function respondToButton(id: string, token: string, text?: string) {
+	const route = Routes.interactionCallback(id, token);
+	if (text) {
+		return globalClient.mahojiClient.restManager.post(route, {
+			body: {
+				type: InteractionResponseType.DeferredMessageUpdate,
+				data: { content: text, flags: MessageFlags.Ephemeral }
+			}
+		});
+	}
+
+	await globalClient.mahojiClient.restManager.post(route, {
+		body: {
+			type: InteractionResponseType.DeferredMessageUpdate
+		}
+	});
+}
